refactored and started making a testing environment for influx.
