+++ app/API/playlist.py
@@ -1,10 +1,4 @@
 from flask_restplus import Namespace, Resource, fields
-from app.utils.tasks import find_song_recommendations
-from app.utils import influx, models
-from app import app
-
-import dateparser
-import datetime
 
 api = Namespace('playlist', description='playlist', path="/playlist")
 
@@ -12,12 +6,11 @@ user_data = api.model("inserted_data", {
     "songid": fields.String,
     "excitedness": fields.Float,
     "happiness": fields.Float
-
 })
 
 
 @api.route('/list/')
-class Song(Resource):
+class Playlist(Resource):
     """Add feedback for song"""
 
     def get(self,userid):
