+++ app/API/song_calls.py
@@ -1,10 +1,5 @@
 from flask_restplus import Namespace, Resource, fields
-from app.utils.tasks import find_song_recommendations
-from app.utils import influx, models
-from app import app
-
-import dateparser
-import datetime
+from app.utils import models
 
 api = Namespace('songs', description='Song information', path="/songs")
 
@@ -12,13 +7,12 @@ user_data = api.model("inserted_data", {
     "songid": fields.String,
     "excitedness": fields.Float,
     "happiness": fields.Float
-
 })
 
 
 @api.route('/mood/<string:userid>')
-class Playlist(Resource):
-    """Add feedback for song"""
+class SongFeedback(Resource):
+    """ Add feedback for song. """
 
     @api.expect(user_data)
     def post(self):
