+++ app/tests/influxdb.conf.template
@@ -1,0 +1,29 @@
+bind-address = ":{global_port}"
+
+[meta]
+  dir = "{meta_dir}"
+  hostname = "localhost"
+  bind-address = ":{meta_port}"
+
+[data]
+  dir = "{data_dir}"
+  wal-dir = "{wal_dir}"
+  index-version = "tsi1"
+
+[http]
+  enabled = true
+  bind-address = ":{http_port}"
+  auth-enabled = false
+
+[[udp]]
+  enabled = {udp_enabled}
+  bind-address = ":{udp_port}"
+  database = "db"
+  batch-size = 0
+  batch-timeout = "0"
+
+[monitor]
+  store-enabled = false
+
+[hinted-handoff]
+  dir = "{handoff_dir}"
