+++ retrain_worker.py
@@ -9,23 +9,23 @@ def main():
     features = ["response_excitedness", "response_happiness", "mode", "time_signature", "acousticness",
                 "danceability", "energy", "instrumentalness", "liveness", "loudness", "speechiness", "valence", "tempo"]
     data = link_features_mood(get_responses=True)
-    trainset = []
+    train_set = []
     print("test")
     print("Number of songs:" + str(len(data)))
     for song in data:
         print(data)
         row = [song[feature] for feature in features]
-        trainset.append(row)
-    trainset = np.array(trainset).astype(float)
-    energy = [elem[1] for elem in trainset]
-    happiness = [elem[2] for elem in trainset]
-    traindata = [elem[5:] for elem in trainset]
-    E_est = GBR(n_estimators=50, max_depth=3)
-    E_est.fit(traindata, energy)
-    H_est = GBR(n_estimators=50, max_depth=3)
-    H_est.fit(traindata, happiness)
-    dump(E_est, 'Retrained-Energy.joblib')
-    dump(H_est, 'Retrained-Happiness.joblib')
+        train_set.append(row)
+    train_set = np.array(train_set).astype(float)
+    energy = [elem[1] for elem in train_set]
+    happiness = [elem[2] for elem in train_set]
+    train_data = [elem[5:] for elem in train_set]
+    excited_est = GBR(n_estimators=50, max_depth=3)
+    excited_est.fit(train_data, energy)
+    happy_est = GBR(n_estimators=50, max_depth=3)
+    happy_est.fit(train_data, happiness)
+    dump(excited_est, 'Retrained-Energy.joblib')
+    dump(happy_est, 'Retrained-Happiness.joblib')
 
 
 if __name__ == "__main__":
