+++ run.py
@@ -1,4 +1,5 @@
 #!flask/bin/python
 from app import app
 
-app.run(host='0.0.0.0', port=5000, threaded=True)
+if __name__ == "__main__":
+    app.run(host='0.0.0.0', port=5000, threaded=True)
